Update Report.md
Original text:
"This project explores how weather of some place at a given time can be predicted by values of weather indicators or by simply an image at a particular time. Possible similarities or differences in implementation or results, between the two approaches will be discussed. In specific, data about a same period of time (09/01/2017 - 10/31/2017) is taken as the experimental input as we want to compare the results of weather prediction in this period, while data prior to 09/01/2017 is used for training the model so that we could know which approach would perform better under the same initial circumstances."
Revised to:
"This project explores how weather can be predicted in a given time by values of weather indicators or by simply an image at a particular time. The implementations or results between those two approaches will be compared with data in a specific period of time (09/01/2017 - 10/31/2017) as the experimental input. Since data prior to 09/01/2017 is used for training the model, the better approach will be found out under the same initial circumstances."