Update Report.md
Original text:
"Predicting the weather is not an easy job, at least it is not enough to make it work by simply putting some data in a simplified ML model. I did not combine the weather data features with the images features together to make predictions because I think it would be  confusing for comparison. But if I had more time I will certainly get this done."
Revised to:
"Simply putting some data in a simplified ML model is not precise enough to predict the weather. The weather data features did not combine with the images features because it will cause confusion in comparison. If the time is allowed, this combination must be done."